backend/requirements.txt
fastapi>=0.95.0
starlette>=0.27.0
httpx>=0.23.0
uvicorn==0.24.0
python-dotenv==1.0.0
langgraph>=0.4.8
langchain>=0.1.0
openai>=1.10.0,<2.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-multipart==0.0.6
requests==2.32.3
pytz>=2023.3
dateparser>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
